% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageFilter.R
\name{cutoff_filter}
\alias{cutoff_filter}
\title{Cutoff filter for RGB vector}
\usage{
cutoff_filter(rgb_vector, threshold = 127)
}
\arguments{
\item{rgb_vector}{A raw vector of 3 RGB values}

\item{threshold}{An integer from 0 to 255 representing the grayscale cut-off value.
Default is set to 127 (approximately 50%).}
}
\value{
A vector with numeric values. If grayscale value is bigger than threshold, returns black (0, 0, 0),
        otherwise returns magenta (255, 0, 255).
}
\description{
Cutoff filter for RGB vector
}
\examples{
cutoff_filter(as.raw(c(200, 100, 150)), threshold = 100) #returns black
}
