% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageFilter.R
\name{greyscale_filter}
\alias{greyscale_filter}
\title{Greyscale filter for RGB vector}
\usage{
greyscale_filter(rgb_vector)
}
\arguments{
\item{rgb_vector}{A raw vector of 3 RGB values}
}
\value{
Function returns numeric vector of greyscale values
}
\description{
Greyscale filter for RGB vector
}
\examples{
greyscale_filter(as.raw(c(210, 55, 22)))
}
