% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageFilter.R
\name{image_filter}
\alias{image_filter}
\title{Immage filter - apply a filter function to each pixel af a bitmap image}
\usage{
image_filter(bitmap_name, filter_name)
}
\arguments{
\item{bitmap_name}{A 3D raw array representing a bitmap image (from image_read()[[1]] subsetting the first element)}

\item{filter_name}{A function that takes an RGB vector (length 3) and returns a transformed RGB vector}
}
\value{
A 3D raw array representing the new filtered bitmap image
}
\description{
Immage filter - apply a filter function to each pixel af a bitmap image
}
